intruder-detection/requirements.txt
attrs
numpy
paho_mqtt
Pillow
ultralytics
